refactor special dist
